Implement modularity. Normalize file structure. Update test commands
