+++ index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+var bitmapReader = require('./lib/bitmapReader');
+
+var app = function() {
+  bitmapReader();
+};
+
+app();
+
+module.exports = app;
