+++ package.json
@@ -4,7 +4,8 @@
   "description": "A conversion tool for transforming .bmp files",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "./node_modules/grunt-cli/bin/grunt test",
+    "start": "node index.js"
   },
   "repository": {
     "type": "git",
